Extract delivery_hook stuff into new file
