+++ crates/common/src/config/smtp/session.rs
@@ -21,13 +21,15 @@ use smtp_proto::*;
 use utils::config::{Config, utils::ParseValue};
 
 use crate::{
-    config::CONNECTION_VARS,
+    config::{CONNECTION_VARS, smtp::delivery_hook::DeliveryHook},
     expr::{if_block::IfBlock, tokenizer::TokenMap, *},
 };
 
 use self::resolver::Policy;
 
 use super::*;
+
+use delivery_hook::parse_delivery_hooks;
 
 #[derive(Clone)]
 pub struct SessionConfig {
@@ -177,18 +179,6 @@ pub struct MTAHook {
     pub tls_allow_invalid_certs: bool,
     pub tempfail_on_error: bool,
     pub run_on_stage: AHashSet<Stage>,
-    pub max_response_size: usize,
-}
-
-#[derive(Clone)]
-pub struct DeliveryHook {
-    pub enable: IfBlock,
-    pub id: String,
-    pub url: String,
-    pub timeout: Duration,
-    pub headers: HeaderMap,
-    pub tls_allow_invalid_certs: bool,
-    pub tempfail_on_error: bool,
     pub max_response_size: usize,
 }
 
@@ -600,79 +590,6 @@ fn parse_hooks(config: &mut Config, id: &str, token_map: &TokenMap) -> Option<MT
         max_response_size: config
             .property_or_default(
                 ("session.hook", id, "options.max-response-size"),
-                "52428800",
-            )
-            .unwrap_or(52428800),
-        headers,
-    })
-}
-
-fn parse_delivery_hooks(config: &mut Config, id: &str, token_map: &TokenMap) -> Option<DeliveryHook> {
-    let mut headers = HeaderMap::new();
-
-    for (header, value) in config
-        .values(("session.delivery_hook", id, "headers"))
-        .map(|(_, v)| {
-            if let Some((k, v)) = v.split_once(':') {
-                Ok((
-                    HeaderName::from_str(k.trim()).map_err(|err| {
-                        format!(
-                            "Invalid header found in property \"session.delivery_hook.{id}.headers\": {err}",
-                        )
-                    })?,
-                    HeaderValue::from_str(v.trim()).map_err(|err| {
-                        format!(
-                            "Invalid header found in property \"session.delivery_hook.{id}.headers\": {err}",
-                        )
-                    })?,
-                ))
-            } else {
-                Err(format!(
-                    "Invalid header found in property \"session.delivery_hook.{id}.headers\": {v}",
-                ))
-            }
-        })
-        .collect::<Result<Vec<(HeaderName, HeaderValue)>, String>>()
-        .map_err(|e| config.new_parse_error(("session.delivery_hook", id, "headers"), e))
-        .unwrap_or_default()
-    {
-        headers.insert(header, value);
-    }
-
-    headers.insert(CONTENT_TYPE, "application/json".parse().unwrap());
-    if let (Some(name), Some(secret)) = (
-        config.value(("session.delivery_hook", id, "auth.username")),
-        config.value(("session.delivery_hook", id, "auth.secret")),
-    ) {
-        headers.insert(
-            AUTHORIZATION,
-            format!("Basic {}", STANDARD.encode(format!("{}:{}", name, secret)))
-                .parse()
-                .unwrap(),
-        );
-    }
-
-    Some(DeliveryHook {
-        enable: IfBlock::try_parse(config, ("session.delivery_hook", id, "enable"), token_map)
-            .unwrap_or_else(|| {
-                IfBlock::new::<()>(format!("delivery.hook.{id}.enable"), [], "false")
-            }),
-        id: id.to_string(),
-        url: config
-            .value_require(("session.delivery_hook", id, "url"))?
-            .to_string(),
-        timeout: config
-            .property_or_default(("session.delivery_hook", id, "timeout"), "30s")
-            .unwrap_or_else(|| Duration::from_secs(30)),
-        tls_allow_invalid_certs: config
-            .property_or_default(("session.delivery_hook", id, "allow-invalid-certs"), "false")
-            .unwrap_or_default(),
-        tempfail_on_error: config
-            .property_or_default(("session.delivery_hook", id, "options.tempfail-on-error"), "true")
-            .unwrap_or(true),
-        max_response_size: config
-            .property_or_default(
-                ("session.delivery_hook", id, "options.max-response-size"),
                 "52428800",
             )
             .unwrap_or(52428800),
